+++ MedTD/Assets/Scripts/Player/Towers/MeleeTowerLevel.cs
@@ -3,8 +3,8 @@
 public class MeleeTowerLevel : TowerLevel
 {
     public float unitSpeed = 1f;
-    public int unitHealth = 10;
-    public int unitDamage = 10;
+    public float unitHealth = 10;
+    public float unitDamage = 10;
     public int unitDefense = 10;
     public float unitHitCooldown = 1f;
     public float meleeHitRange = 1f;
